graphql: fix websocket autoReconnect on server disconnect
Previously cancelOnError was set to true for the _messages stream, so
onConnectionLost would not get called
after an error (such as a
WebSocketChannelException on a forced disconnect from the server), and
thus it would not attempt to
reconnect.

This was manually tested on the web platform to reconnect after the
proper duration after the server forcefully disconnects a websocket.