feat(graphql): add possibility to update the client components
This is a very opinionated API, but it is required because it hide the complexity of the client under the hood.

The user should never know the how we use the component and how the client it built.

The user need to know that we are like Harry Potter and the stuff happens like magic.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>